+++ src/cm_kind.mli
@@ -1,9 +1,13 @@
+open Stdune
+
 type t = Cmi | Cmo | Cmx
 
 val all : t list
 
 val ext : t -> string
 val source : t -> Ml_kind.t
+
+val to_sexp : t Sexp.To_sexp.t
 
 module Dict : sig
   type cm_kind = t