+++ CHANGES.md
@@ -38,6 +38,10 @@ next
 
 - Improve message suggesting to remove parentheses (#1196, fix #1173, @emillon)
 
+- Add `(wrapped (transition "..message.."))` as an option that will generate
+  wrapped modules but keep unwrapped modules with a deprecation message to
+  preserve compatibility. (#1188, fix #985, @rgrinberg)
+
 1.1.1 (08/08/2018)
 ------------------